Added GetCurrentAnimationParametersQuery.

+++ RaspberryLights.Domain/Models/AnimationParameters.cs
@@ -7,7 +7,7 @@ public class AnimationParameters
     public Animation Animation { get; set; }
     public SimpleColor CustomColor { get; set; }
     public SpeedType SpeedType { get; set; }
-    public int UserDefinedSpeed { get; set; }
+    public byte UserDefinedSpeed { get; set; }
     public byte Brightness { get; set; }
 
     public class SimpleColor
